weather api, forget password, social login, gurbara without auth_token, add conditon to show video in filmy, make changes according to api in my profile, update profile and update profile image, change in registration vary of social login or not

+++ app/src/main/java/com/knovatik/navadesh/model/weather/Location.kt
@@ -4,7 +4,7 @@ data class Location(
     val country: String,
     val lat: Double,
     val localtime: String,
-    val localtime_epoch: Int,
+    val localtime_epoch: Long,
     val lon: Double,
     val name: String,
     val region: String,
